Move $.Parsers' implementation to a new $.Parsers_Impl unit
Recent developments in Libadalang/GNAT increased the number of such
symbols, and now Libadalang shared libraries cannot be built anymore on
Windows because the DLL file format limit for exported symbols has been
reached.

The goal of this refactoring is to reduce the number of exported symbols
for shared libraries: unlike $.Parsers, $.Parsers_Impl can be excluded
from the list of "interfaces" at the GPR level, and so its symbols do
not need to be exported.