Unparsing: refactor probing of template category
There are currently two kinds of template: the ones that contain
"recurse", and the ones that contain "recurse_field"/"text". Due to the
current template format, the only way to determine what is the kind of a
given template is to look for the content of that template.

Currently, this is done in the middle of the single pass JSON templates
parsing. However, upcoming work will add another kind of templates:
mixes of "recurse"/"text" for field templates. In order to make this
machinery tractable with this new case, we need to have the detection
working on the JSON as a first pass, and the translation from JSON to
our internal representation as a second pass (so that
translation/validation can rely on the detected kind from the start).

This commit splits template parsing into these two passes, as
preliminary work to support text in field templates.