Unparsing: refactor field unparsing
Move the handling of pre/post tokens to the Unparse_Field subprogram, so
neither Iterate_On_Fragments nor special handling for With_Recurse_Field
node templates have to deal with them manually.

This refactoring also serves as preparatory work for the handling of
"text" nodes in field templates: it will be the responsibility of field
templates to deal with pre/post tokens.