Unparsing: refactor the handling of recurse_flatten nodes
Instead of storing the original node directly inside a document, keep
track of this information locally during template instantiation.

The design is clearer overall: it was strange to have a Node component
in Document_Type that was sometimes null, sometimes with an actual node,
which comes from the fact that there is no 1:1 mapping between nodes and
document nodes, and some documents are created in contexts where it is
not obvious for which node they are created.

This is just a refactoring: no change of behavior intended.