Lkt lowering: fix token forward references in lexer match rules
The order of rules in lexers is not meant to be significant. Since
"match" rules can refer to tokens, this means that it must be possible
to refer to a given token from a "match" rule even if that token is
defined later in the source code.

Lkt lowering was not supporting this so far: this commit plugs this
hole.