Lkt: fix dynamic variable binding when parent scope has already bound it
To achieve this, introduce a new type of scope entity (BoundDynVar) to
materialize the fact that a dynamic variable has been bound. We
previously used LocalVar entities instead to materialize this, which was
hiding the fact that this was about a dynamic variable, and thus
preventing it from being rebound.