AbstractVariable: assert creating of local var happens before construct
Creating a local variable for an AbstractVariable may change the name of
that variable (to avoid naming conflicts in the generated code). So it
has to happen before constructing the corresponding VariableExpr (so
that this VariableExpr refers to the final name).

Because the Lkt lowering pass generates different names than equivalent
DSL expressions, it used to trigger a name-changing local variable
generation, which caused code generation problems, visible as an "unused
binding" warning during code emission.