Lexical envs: fix logging involving the empty environment
Now that lexical env handling code uses thin symbols, a reference to the
analysis context symbol table is needed to compute the image of a
symbol. All lexical envs do store a reference to this symbol table,
except one: the empty env. Instead of adding a reference to the symbol
table to all subprograms involved or creating one empty env per context,
just add defensive code to avoid the crash when the symbol table is
missing. Operations on the empty environment are generally
uninteresting, so it should not be a big deal not to have the symbol
name in this case.