Remove dependency to Camomile
Use the newly introduced utf8/utf32 converters.

This makes some changes on the memory management as unwrapping a
TextType is now handled by an utf8/utf32 converter which allocates
the memory for us. This means that destroy_text is now required as a
finalizer for a Text type.

Issue eng/codepeer/codepeer#1185