Unparsing: remove ifBreak nodes when the group is known to be broken
This simplifies the tree, and in particular removes a disjunction, and
thus avoids annoying spacing requirement enforcing issues. For instance:

  [ifBreak([token1, hardlineWithoutBreakParent], token1), token2]

If a hardline is needed between token1 and token2, then the
Insert_Required_Spacing will append a hardline before token2, which will
be duplicated with the one in the "break" branch of ifBreak.

If we know that the parent group is broken, we know we can replace the
above with:

  [[token1, hardlineWithoutBreakParent], token2]

And thus there is no need to insert a hardline.