Unparsing: eagerly compute unparsing fragments for trivias
So far, unparsing fragments for trivias were computed in the middle of
the unparsing of the tree (that is, as surrounding tokens were
processed).

The upcoming addition of support for Prettier tables will need to have
access to these unparsing fragments (to determine when to split tables).
To achieve that, it is simpler to compute all unparsing fragments
for trivias as a preliminary pass, and then let the tree unparsing and
tables processing query these unparsing fragments.

In addition, this will make it straightforward to ignore surrounding
trivias when unparsing a node that is not the root of its unit (just
iterate through the tokens for that node).

This commits reworks the unparsing engine to create this preliminary
pass.