Lkt lowering: lower lexing rules in the expected order
The order for lexing rules is significant (for a same input length, the
first lexing rule that matches has precedence). To lower them in the
right order and still be able to deal with forward references, lower
lexers in two passes: first handle token declarations, and only then
lower rules.