Unparsing: fix the handling of comments that immedialely precede groups
When running the unparser for code patterns such as:

    # Comment
    foo bar

Then it is necessary to insert a line break between the comment an what
comes after. This is generally done by the "insert spacing" pass, which
inserts spacing right before token T2 if special spacing is required
between T1 and T2.

In the case above, this gives unwanted results: if "foo bar" is enclosed
in a group, then this means that the line break will be inserted inside
that group, and thus the group will be broken (because it will contain a
hard line break). As a consequence, the final formatting will turn the
"line" Prettier document between "foo" and "bar" into a line break.

Tweak the trivias sequence analysis pass to insert a line break right
after the comment (i.e. before potential group documents that come
after), so that such groups are not broken because of comments.