Lkt: make it possible to create the unit provider in the Python API
On Windows, attempts to modify the environment variables from Python
(os.environ) will not affect the environment variables reachable from
Ada/C. As a result, attempts to modify the LKT_PATH environment variable
from the manage.py script will not allow the Lkt lowering pass to find
Lkt source files.

To fix this, extend the Python bindings of Liblktlang to allow the
creation of a unit provider from the LKT_PATH environment variable found
in os.environ, and update the Lkt lowering pass to use that.