Unparsing: fix handling of reattached trivias for "wrapper" nodes
When a list node (A) child (B) has the same "token end" as its own child
(C), reattached tokens used to be processed when unparsing the last
token of (C), whereas they are intended to be processed as "inter-child"
trivias during the unparsing of (A). As a consequence, trivias that
appear between two list children could be unparsed inside prettier
constructs that belong to individual elementrs, resulting in spuriously
indented comments and broken groups.

Enhance the reattached trivias tracking data structure so that we
control more finely when reattached tokens are processed to avoid this
kind of problem.