parsers/list_code_ada.mako: fix token_end when not allow_trailing
Unless allow_trailing is present for a list parser, any token that
matches a list separator and that follows the list node (i.e. when that
separator is not followed by a list element) must be excluded from the
token range for this lis node. See the test addition for an example.