[ocaml api] Avoid calling dec_ref for array items
Calling dec_ref on an array item is completely wrong as the dec_ref
function of the array itself is already calling each items dec_ref. To
avoid doing so, pass an additional parameter 'dec_ref' to the wrap
functions of struct and arrays and only call dec_ref if this value is
true.

Closes eng/libadalang/langkit#831