dsl_unparse.py: fix locals/globals precedence in template evaluation
Local variables are supposed to mask global variables, not the other way
around, so the correct way to compute the set of variables for template
evaluation is `dict(globals, **locals)`.

Future work will add `from __future__ import annotations` to
dsl_unparse.py, which will reveal this bug as the `annotations` feature
(a global) will hide the `annotations` function (a local).