Refactor computation of main programs
Let manage.py scripts give the set of extra mains (instead of computing
the whole set), and delay the computation of all mains to the emit step:
this needs to know if unparsers are enabled.

Future work will introduce a grammar annotation to enable unparsers, so
computing the set of mains will be possible only once the grammar has
been lowered.