langkit.expressions.logic: remove mistakenly added line
During a resolution of a rebase conflict, a line was mistakenly added,
causing property expressions to be constructed multiple times, and
eventually to have local variables created for nothing, thus reported as
unused by GNAT. Remove it.

Note that this was not caught by our testsuite because warnings are
disabled for generated test libraries: we thought there was too much
special code to add to avoid all the compiler warnings currently emitted
for test (dummy) libraries.