Unparsing: do not reattach tokens to "absent" empty lists
When an empty list is produced by the "no match" case for an "opt"
parser for a field, then the unparsing engine considers that that field
(initialized by the empty list) is "absent", so it will be discarded by
the unparsing algorithm. This means that tokens which are reattached to
"absent" empty lists will not be in fact processed: they will be lost
at the end of the unparsing process.

To avoid this problem, and because it also yields a more intuitive
process, never reattach tokens to such empty lists.