Unparsing: introduce a batch mode for the unparsing main
This will make it possible to run the auto-checks on a list of files in
a single run directly rather than running the unparsing main once per
file to reformat. This is an optimization that allows to save the time
it takes to spawn a subprocess, create an analysis context and load the
unparsing configuration.