Testsuite: run Langkit scripts through "python -m langkit.scripts.*"
On Windows, the executables created by "pip install" hardcode the
absolute path to a Python interpreter. On production servers, the Python
interpreter (whose path is determined during the build of the Langkit
package) are located elsewhere during testing, so in practice these
executables do not work. Use the "python -m langkit.scripts.*" form to
run these scripts instead, which is more robust.