Testsuite: add missing calls to sys.stdout.flush() for consistent output
Flushing is necessary whenever there is potentially buffered content in
Python's sys.stdout and the test is about to run native code (call to
a C/Ada routine, or spawning a subprocess). The couple of adjustments in
this commit were found necessary with testing on a local Windows machine
(not visible in production).