langkit.libmanage: remove the top-level Exception handler
This exception handler was needed to catch errors in the Python DSL,
which could materalize as any kind of exception. Now that the Python DSL
is retired, all DSL errors are supposed to be reported directly using
langkit.diagnostics, so this wide handler can only catch Langkit
bugs/crashes, which hinders debugging.