Lkt lowering: process default field values during EXPR_LOWERING
Do not lower expressions for field default values during types lowering,
as lowering the expression may need access to not-yet-lowered types:
this is not an issue right now as type references are currently turned
into `TypeRepo.Defer` instances, to be resolved later, but will create
problems with upcoming changes. Do this lowering during the dedicated
expression lowering pass instead, to avoid this problem.