Refactor processings for AbstractNodeData names
Create a dataclass dedicated to centralize the definition and creation
logic of names for all the kinds of "members" (`AbstractNodeData`
subclasses). This is an attempt at clarifying and harmonizing the
semantics of each kind of name, to ease future maintenance in this area.

The `properties/internal_use` testcase is removed because, thanks to the
harmonized logic, it is no longer technically possible to access an
internal member from the language specification, so it is not possible
to get an error when trying to access one.