Determine compiled type member owner at member initialization time
With the Python DSL, `AbstractNodeData` instances has to be created
before their owning compiled type, so the backlink to their owner had to
be initialized later. Now that the Python DSL is gone and thanks to the
previous work that deferred the creation of members after all types are
known, this is no longer needed: add an `owner` argument to the
constructors of member classes so that this backlink is set as early as
possible.

Also rename the `struct` attribute to `owner`, to better reflect the
fact that all compiled types can have members (this used to be reserved
to structs and nodes only).