Make type information mandatory for all members
Now that type members are created after all compiled types are
available, and that the type of members is no longer inferred (from
inheritted members or from the grammar: the type annotation is
mandatory), it is no longer necessary for `AbstractNodeData` and its
subclasses to accept None or a `TypeRepo.Defer` value as their type at
instantiation time.

Update all type signatures accordingly and remove obsolete type
inference code.