langkit.frontend.resolver: do not use TypeRepo.Defer
The `error_location` builtin dynamic variable has the root node type.
Since the creation of builtin entities happen before the root node type
is found/lowered, we need to use the deferring mechanism so that its
type is assigned only once the root node type is known (our goal is to
get rid of `TypeRepo.Defer`).