langkit.expressions: assume only one property at a time is compiled
In the days of the Python DSL, property return types were inferred from
their bodies, and so the compilation of one property could trigger the
compilation of another property, so supporting recursive compilation was
necessary.

In Lkt, property return types must be explicit in the source code, so
recursive compilation is no longer needed: simplify compilation code
accordingly.