Stop using AbstractVariable to represent local variables internally
Extend `langkit.expressions.base.LocalVars.LocalVar` so that it
can represent manully declared entities and use that of
`AbstractVariable` for this purpose.

After this commit, all `AbstractVariable` are thin wrapper around
`LocalVar` to be used as local variable references in abstract
expressions, and have `VariableExpr` (references in resolved
expressions) contain an optional reference to `LocalVar` (if it
translates a reference to a local variable).