properties/invalid_then: fix an error in the grammar
This error went unnoticed so far because another fatal compilation error
was hiding it. Upcoming refactorings will change the order for the
corresponding checks. Fix this error as it is not part of what this test
is meant to exercise, so that the test result does not depend on the
order of checks in Langkit.