langkit.utils: remove inherited_property
As the ongoing work to add type hints to the rest of the codebase shows,
this function is just too delicate to annotate correctly. In addition,
the inheritance logic is actually dead code, as the only two use cases
(snaps and rebindable node annotations) assigns non-None values for each
node, so these properties are never inherited in practice.