Lkt: rework collection subscript expressions
Introduce `a.at(i)` and `a?.at(i)` as out-of-bounds resilient
alternatives to `a[i]` and `a?[i]`. In addition, harmonize the effect of
`?` on there expressions with its effect on member access (`a?.b`): if
the prefix is null, the execution of the rest of the chain is
shortcircuited.