Generic API: add bindings for the rewriting features
Move the implementation of the rewriting engine to Langkit_Support and
add bindings to it in the generic API. Rework the C bindings to be
implemented on top of the generic API.

Note that no new testcase is added: the generic API is tested indirectly
by all existing rewriting testcases by the virtue of being the
common implementation behind all the other rewriting APIs. The Ada API
testcase will transition to the generic API once the Ada API is removed.