Generic API: use language-agnostic ways to fetch context version numbers
Instead of relying on a language descriptor operation, use the fact that
the `Analysis_Context_Type.Serial_Number` record component has a
consistent ABI in all generated libraries to fetch the version number of
contexts directly. This is supposedly more efficient, and more
importantly allows us to perform safety net checks even without a
reference to the language descriptor available. This will be useful for
the upcoming work of bringing the rewriting API to
`Langkit_Support.Generic_API`.