Add validity checks for predicate error message templates
Add decent error messages for two error cases: when the template
references a non-existing argument, and when the referenced argument is
not an entity.

Since error messages must be emitted before code emission, this commit
also reorganizes template handling logic so that a first part runs
during Lkt lowering (this is where validation checks are done) and the
second part runs during code generation (that part relies on validity
checks to always succeed in generating code).