Rewriting: determine which nodes should be reformatted
This commit adds datastructures to determine which parts of the
rewriting tree must be reformatted: the node rewriting handles are
partitionned into "rewriting tiles", so that all nodes that belong to
the same tile are either all reformatted, or not at all (i.e. their
formatting in the rewritten source is copied from the original source).

Note that this does not implement partial reformatting itself: this is
just preparatory work.