Rework the emission of "unused bindings" warnings
Stop basing the emission of these warnings on the analysis of the
expression tree emitted for a given property. Use instead the Scope data
structure in the Lkt frontend to track which entities are looked up
(i.e. used somewhere during the translation of Lkt to expressions).

This allows more precise warnings for the cases where a single lowered
variable is used to materialize two different Lkt-level variables (for
instance in `.[i]filtermap` constructs), and in other similar cases.