langkit.emitter: lazily load the default unparsing configuration
Do not load the default unparsing configuration when instantiating the
``Emitter`` class, as it is used only when actually generating the code.
That class is instantiated even when just building the generated
project, so a missing default unparsing config file should not be a
problem in this context.

This will allow us to remove the default unparsing config file from
bootstrap sources.