Lkt: add error recovery to the parsing of block expression
Introduce a new node (`BlockExprClause`) so that:

1. The `;` token belongs to list items in block expressions rather than
   being a separator for that list.

2. The inner expression in block exprs belongs to that list, so that
   there is no `;` token parser that follows the parser for that list.

The conjunction of both points allows us to introduce `skip`/`dont_skip`
parser for error recovery in the parsing of block expressions. The extra
sources that are accepted by the new parser are rejected by the Lkt
frontend.

Note that, due to a Langkit parser bug (likely
eng/libadalang/langkit#652), sometimes there is no parsing error even
though the parse tree contains an error node. This should be addressed
later.