Rewriting: do not load an unparsing configuration by default
When no unparsing configuration is passed to start a rewriting session,
do not try to load the default unparsing configuration (which is costly)
and preserve previous formatting anyway.

This improves the performance of rewriting API users like instrumenters
(which do not care about formatting) which may create one rewriting
session per source file to instrument.