Parsers: include Fail_Info in packrat tables
Having the `Fail_Info` component of the `Parser_Type` data type updated
is instrumental to correctly implementing error recovery.

When storing a parsing failure in a packrat table, keep track of the
`Fail_Info` data structure so that it can be restored when reusing that
parsing result, i.e. so that the effect of calling a parser function on
the `Parser.Last_Fail` component is the same whether the parser actually
runs or returns a result from a packrat table.