Parsers: simplify code generation
Instantiate `Langkit_Support.Packrat.Tables` only once (for the root
node type): all node types are access to the root node type, the only
thing that changes is the associated dynamic predicate, which brings no
value in parsers in practice.

This reduces the amount of generated code, and allows us to have a
single set of types for packrat tables, which will in turn allow future
changes to have a single type for parsing failure information.