$.Implementation: avoid implementation-dependent evaluation order
Ada does not specify the order of evaluation for call actuals. As a
consequence, in `Dump_Lexical_Env`, `Get_Env_Id` may currently be called
first on the parent env, or on the current env depending on what the
compiler thinks is best for code generation, and so the order in which
IDs are assigned may change from one platform to another or from one
compiler version to another.

Rework calls to `Get_Env_Id` so that the order is stable.