Lkt frontend: use scopes to resolve base type references
Remove the `LktTypesLoader.named_types` attribute, and use instead
regular scopes to resolve base types when lowering node types. Since
this happens in the middle of the type lowering pass, for this to work,
scopes need to contain placeholders for types (i.e. `UserType` entities
with no associated `CompiledType` instance).

Upcoming work (Lkt modules) will make entity resolution more complex, so
it is best to rely solely on scopes (which will be aware of Lkt modules)
to perform entity resolution.