[ocaml api] fix correctly referencing analysis context in token
Tokens had a 'fake' reference to the analysis context that the OCaml GC
couldn't really follow. Which means that if there were no real other
references to the analysis context that created the token, further uses
of the token could crash the application.

Fix this by splitting the Token module into:
 - TokenStruct for the low level bindings to the c token structure
 - Token for the high level view of a token

The high level view is what the user of the API is going to manipulate
and is containing a real reference to the AnalysisContext that the OCaml
GC is able to register.

Closes eng/libadalang/langkit#987